1. Add array labels (label='foo[3]') 2. Put two psd algorithms together by using boost adaptors and templates 3. Add aperture deleters to Roro

+++ pkg/dem/ParticleContainer.cpp
@@ -22,7 +22,7 @@ void ParticleContainer::clear(){
 }
 
 Particle::id_t ParticleContainer::findFreeId(){
-	size_t size=parts.size();
+	long size=(long)parts.size();
 	auto idI=freeIds.begin();
 	while(!freeIds.empty()){
 		id_t id=(*idI); assert(id>=0);
